refactor : Folder -> shared 페이지로 변경
