+++ src/App.js
@@ -3,6 +3,7 @@ import Nav from './components/Nav/Nav';
 import GlobalStyle from './styles/GlobalStyle';
 import { Route, Routes } from 'react-router-dom';
 import Folder from './pages/Folder';
+import Shared from './pages/Shared';
 
 
 function App() {
@@ -12,6 +13,7 @@ function App() {
       <Nav />
       <Routes>
         <Route path='/folder' element={<Folder />} />
+        <Route path='/shared' element={<Shared />} />
       </Routes>
       <Footer />
     </div>
